extracting parse fail tests
I'd expect pa_modargs_get_value_u32() to return an error code when extracting fails. However,
it leaves the value unchanged and returns success. Maybe that is intended, like that the
value is used as both input for the default value and output for the result, but then the
function isn't well documented. In particular, it means that you must init your variable
before calling instead of using a default value if extracting fails.